+++ src/BaseHhProvider.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Andy87\ClientsHh;
+namespace and_y87\ClientsHh;
 
-use Andy87\PhpClientSdk\Provider\AbstractProvider;
+use and_y87\PhpClientSdk\Provider\AbstractProvider;
 
 /**
  * Базовый provider для Hh API с ленивым доступом к вложенным группам методов.